refactor: move get_hpcname to existing hpc module
also improve test coverage
