--- src/ccbr_tools/get_hpcname.py
@@ -1,10 +1,0 @@
-from ccbr_tools.pipeline.hpc import get_hpcname
-
-
-def main():
-    HPC = get_hpcname()
-    print(f"{HPC}")
-
-
-if __name__ == "__main__":
-    main()
